+++ docs/source/changelog.rst
@@ -27,6 +27,7 @@ Changelog
   - Add a function to load vertical levels data
   - Improve `interp_to_pres_lev()`
   - Add `model` keyword to `plot.pv` functions
+  - Rewrite `DimConstr` API
 
 
 0.4.6